Avoid breaking the storage version

+++ test/extension/test_remote_optimizer.cpp
@@ -77,7 +77,7 @@ TEST_CASE("Test using a remote optimizer pass in case thats important to someone
 
 			auto buffer = malloc(bytes);
 			REQUIRE(buffer);
-			REQUIRE(read(connfd, buffer, bytes) == bytes);
+			REQUIRE(read(connfd, buffer, bytes) == ssize_t(bytes));
 
 			BufferedDeserializer deserializer((data_ptr_t)buffer, bytes);
 			con2.BeginTransaction();
@@ -97,7 +97,7 @@ TEST_CASE("Test using a remote optimizer pass in case thats important to someone
 				auto data = serializer.GetData();
 				idx_t len = data.size;
 				REQUIRE(write(connfd, &len, sizeof(idx_t)) == sizeof(idx_t));
-				REQUIRE(write(connfd, data.data.get(), len) == len);
+				REQUIRE(write(connfd, data.data.get(), len) == ssize_t(len));
 			}
 		}
 		exit(0);
